backend/requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.8.3
click==8.2.1
distro==1.9.0
dnspython==2.7.0
email-validator==2.3.0
fastapi==0.104.0
greenlet==3.2.4
groq==0.4.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.4
httpx==0.25.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2-binary==2.9.10
pydantic==2.7.1
pydantic-settings==2.6.1
pydantic_core==2.18.2
python-dotenv==1.0.0
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.15.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==0.21.0
websockets==15.0.1

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.0
factory-boy==3.3.0
faker==20.1.0